+++ dask_ml/decomposition/incremental_pca.py
@@ -13,7 +13,7 @@ from dask import delayed
 from sklearn.utils import gen_batches
 from sklearn.utils.validation import check_random_state
 from ..utils import check_array
-from ..utils import svd_flip_fixed
+from ..utils import _svd_flip_copy
 from .._utils import draw_seed
 from .extmath import _incremental_mean_and_var
 from . import PCA
@@ -24,7 +24,7 @@ def svd_flip(u, v):
     This is a replicate of svd_flip() which calls svd_flip_fixed()
     instead of skm.svd_flip()
     """
-    u2, v2 = delayed(svd_flip_fixed, nout=2)(u, v, False)
+    u2, v2 = delayed(_svd_flip_copy, nout=2)(u, v, False)
     u = da.from_delayed(u2, shape=u.shape, dtype=u.dtype)
     v = da.from_delayed(v2, shape=v.shape, dtype=v.dtype)
     return u, v
@@ -213,3 +213,54 @@ class IncrementalPCA(PCA):
         else:
             self.noise_variance_ = 0.
         return self
+
+    def get_covariance(self):
+        """Compute data covariance with the generative model.
+        ``cov = components_.T * S**2 * components_ + sigma2 * eye(n_features)``
+        where S**2 contains the explained variances, and sigma2 contains the
+        noise variances.
+        Returns
+        -------
+        cov : array, shape=(n_features, n_features)
+            Estimated covariance of data.
+        """
+        components_ = self.components_
+        exp_var = self.explained_variance_
+        if self.whiten:
+            components_ = components_ * np.sqrt(exp_var[:, np.newaxis])
+        exp_var_diff = np.maximum(exp_var - self.noise_variance_, 0.)
+        cov = np.dot(components_.T * exp_var_diff, components_)
+        cov += np.eye(len(cov)) * self.noise_variance_  # modify diag inplace
+        return cov
+
+    def get_precision(self):
+        """Compute data precision matrix with the generative model.
+        Equals the inverse of the covariance but computed with
+        the matrix inversion lemma for efficiency.
+        Returns
+        -------
+        precision : array, shape=(n_features, n_features)
+            Estimated precision of data.
+        """
+        n_features = self.components_.shape[1]
+
+        # handle corner cases first
+        if self.n_components_ == 0:
+            return np.eye(n_features) / self.noise_variance_
+        if self.n_components_ == n_features:
+            return da.linalg.inv(self.get_covariance())
+
+        # Get precision using matrix inversion lemma
+        components_ = self.components_
+        exp_var = self.explained_variance_
+        if self.whiten:
+            components_ = components_ * np.sqrt(exp_var[:, np.newaxis])
+        exp_var_diff = np.maximum(exp_var - self.noise_variance_, 0.)
+        precision = np.dot(components_, components_.T) / self.noise_variance_
+        precision += np.eye(len(precision)) / exp_var_diff
+        precision = np.dot(components_.T,
+                            np.dot(da.linalg.inv(precision), components_))
+        precision /= -(self.noise_variance_ ** 2)
+        precision += np.eye(len(precision)) / self.noise_variance_
+
+        return precision
