+++ dask_ml/decomposition/pca.py
@@ -13,7 +13,7 @@ from sklearn.utils.validation import check_random_state
 
 from .._compat import check_is_fitted
 from .._utils import draw_seed
-from ..utils import svd_flip_new
+from ..utils import svd_flip
 
 _TYPE_MSG = (
     "Got an unsupported type ({}). Dask-ML's PCA only support Dask Arrays or "
@@ -281,7 +281,7 @@ class PCA(sklearn.decomposition.PCA):
             U, S, V = da.linalg.svd_compressed(
                 X, n_components, n_power_iter=n_power_iter, seed=seed
             )
-        U, V = svd_flip_new(U, V)
+        U, V = svd_flip(U, V)
 
         explained_variance = (S ** 2) / (n_samples - 1)
         components, singular_values = V, S
@@ -504,57 +504,6 @@ class PCA(sklearn.decomposition.PCA):
             Average log-likelihood of the samples under the current model
         """
         return da.mean(self.score_samples(X))
-
-    def get_covariance(self):
-        """Compute data covariance with the generative model.
-        ``cov = components_.T * S**2 * components_ + sigma2 * eye(n_features)``
-        where S**2 contains the explained variances, and sigma2 contains the
-        noise variances.
-        Returns
-        -------
-        cov : array, shape=(n_features, n_features)
-            Estimated covariance of data.
-        """
-        components_ = self.components_
-        exp_var = self.explained_variance_
-        if self.whiten:
-            components_ = components_ * np.sqrt(exp_var[:, np.newaxis])
-        exp_var_diff = np.maximum(exp_var - self.noise_variance_, 0.)
-        cov = np.dot(components_.T * exp_var_diff, components_)
-        cov += np.eye(len(cov)) * self.noise_variance_  # modify diag inplace
-        return cov
-
-    def get_precision(self):
-        """Compute data precision matrix with the generative model.
-        Equals the inverse of the covariance but computed with
-        the matrix inversion lemma for efficiency.
-        Returns
-        -------
-        precision : array, shape=(n_features, n_features)
-            Estimated precision of data.
-        """
-        n_features = self.components_.shape[1]
-
-        # handle corner cases first
-        if self.n_components_ == 0:
-            return np.eye(n_features) / self.noise_variance_
-        if self.n_components_ == n_features:
-            return da.linalg.inv(self.get_covariance())
-
-        # Get precision using matrix inversion lemma
-        components_ = self.components_
-        exp_var = self.explained_variance_
-        if self.whiten:
-            components_ = components_ * np.sqrt(exp_var[:, np.newaxis])
-        exp_var_diff = np.maximum(exp_var - self.noise_variance_, 0.)
-        precision = np.dot(components_, components_.T) / self.noise_variance_
-        precision += np.eye(len(precision)) / exp_var_diff
-        precision = np.dot(components_.T,
-                            np.dot(da.linalg.inv(precision), components_))
-        precision /= -(self.noise_variance_ ** 2)
-        precision += np.eye(len(precision)) / self.noise_variance_
-
-        return precision
 
 
 def _known_shape(shape):
