+++ dask_ml/utils.py
@@ -23,14 +23,14 @@ from ._utils import ConstantFunction
 logger = logging.getLogger()
 
 
-def _svd_flip_copy(x, y):
+def _svd_flip_copy(x, y, u_based_decision=True):
     # If the array is locked, copy the array and transpose it
     # This happens with a very large array > 1TB
     # GH: issue 592
     try:
-        return skm.svd_flip(x, y)
+        return skm.svd_flip(x, y, u_based_decision=u_based_decision)
     except ValueError:
-        return skm.svd_flip(x.copy(), y.copy())
+        return skm.svd_flip(x.copy(), y.copy(), u_based_decision=u_based_decision)
 
 
 def svd_flip(u, v):
